Refactor/imports (#109)
