+++ src/snowbunk/snowbunkClient.ts
@@ -1,7 +1,7 @@
+import DiscordClient from '@/discord/discordClient';
+import userID from '@/discord/userID';
+import webhookService from '@/webhooks/webhookService';
 import { Events, Message, TextChannel } from 'discord.js';
-import DiscordClient from '../discord/discordClient';
-import userID from '../discord/userID';
-import webhookService from '../webhooks/webhookService';
 
 export default class SnowbunkClient extends DiscordClient {
 	private readonly channelMap: Record<string, Array<string>> = {
