+++ src/starbunk/bots/reply-bots/musicCorrectBot.ts
@@ -1,6 +1,6 @@
+import ReplyBot from '@/starbunk/bots/replyBot';
+import { WebhookService } from '@/webhooks/webhookService';
 import { Message, TextChannel } from 'discord.js';
-import { WebhookService } from '../../../webhooks/webhookService';
-import ReplyBot from '../replyBot';
 
 export default class MusicCorrectBot extends ReplyBot {
 	constructor(webhookService: WebhookService) {
