+++ src/starbunk/bots/reply-bots/pickleBot.ts
@@ -1,8 +1,8 @@
+import UserID from '@/discord/userID';
+import ReplyBot from '@/starbunk/bots/replyBot';
+import Random from '@/utils/random';
+import { WebhookService } from '@/webhooks/webhookService';
 import { Message, TextChannel } from 'discord.js';
-import UserID from '../../../discord/userID';
-import Random from '../../../utils/random';
-import { WebhookService } from '../../../webhooks/webhookService';
-import ReplyBot from '../replyBot';
 
 export default class PickleBot extends ReplyBot {
 	constructor(webhookService: WebhookService) {
