+++ src/starbunk/starbunkClient.ts
@@ -1,9 +1,9 @@
+import { Command } from '@/discord/command';
+import DiscordClient from '@/discord/discordClient';
+import { Logger } from '@/services/Logger';
 import { PlayerSubscription } from '@discordjs/voice';
 import { Base, ClientOptions, Collection, Events, Message, REST, Routes, VoiceState } from 'discord.js';
 import { readdirSync } from 'fs';
-import { Command } from '../discord/command';
-import DiscordClient from '../discord/discordClient';
-import { Logger } from '../services/Logger';
 import ReplyBot from './bots/replyBot';
 import VoiceBot from './bots/voiceBot';
 import { DJCova } from './dJCova';
