+++ src/tests/unit/chaosBot.test.ts
@@ -1,7 +1,7 @@
-import { createMockGuildMember, createMockMessage } from '@/test/mocks/discordMocks';
-import { createMockWebhookService } from '@/test/mocks/serviceMocks';
+import ChaosBot from '@/starbunk/bots/reply-bots/chaosBot';
+import { createMockGuildMember, createMockMessage } from '@/tests/mocks/discordMocks';
+import { createMockWebhookService } from '@/tests/mocks/serviceMocks';
 import { Message, User } from 'discord.js';
-import ChaosBot from '../chaosBot';
 
 describe('ChaosBot', () => {
 	let chaosBot: ChaosBot;
