+++ src/tests/unit/gundamBot.test.ts
@@ -1,7 +1,7 @@
-import { createMockGuildMember, createMockMessage } from '@/test/mocks/discordMocks';
-import { createMockWebhookService } from '@/test/mocks/serviceMocks';
+import GundamBot from '@/starbunk/bots/reply-bots/gundamBot';
+import { createMockGuildMember, createMockMessage } from '@/tests/mocks/discordMocks';
+import { createMockWebhookService } from '@/tests/mocks/serviceMocks';
 import { Message, User } from 'discord.js';
-import GundamBot from '../gundamBot';
 
 describe('GundamBot', () => {
 	let gundamBot: GundamBot;
