+++ src/webhooks/webhookService.ts
@@ -1,6 +1,6 @@
+import guildIDs from '@/discord/guildIDs';
+import { MessageInfo } from '@/discord/messageInfo';
 import { Client, Message, TextChannel, Webhook } from 'discord.js';
-import guildIDs from '../discord/guildIDs';
-import { MessageInfo } from '../discord/messageInfo';
 
 export class WebhookService {
 	getWebhookName(channelName: string, isSnowbunk: boolean): string {
